requirements.txt
atomicwrites==1.4.0
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==2.2.9
django-filter==2.3.0
djangorestframework==3.11.1
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
jsonpath==0.82
Markdown==3.2.2
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
zipp==3.2.0
